backend/requirements.txt
# Core Framework
fastapi==0.119.1
uvicorn[standard]==0.38.0
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.8.2

# Google Gemini & LangChain
langchain==1.0.2
langchain-core==1.0.0
langchain-google-genai==2.0.8
google-generativeai==0.8.3

# LangGraph (optional - if using workflows)
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37

# Environment & Configuration
python-dotenv==1.0.1

# HTTP & Networking
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
requests-toolbelt==1.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Utilities
tenacity==9.1.2
PyYAML==6.0.3
orjson==3.11.3
click==8.3.0

# Dependencies
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
ormsgpack==1.11.0
packaging==25.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing-extensions==4.15.0
urllib3==2.5.0
xxhash==3.6.0
zstandard==0.25.0
